requirements.txt
kafka-python
streamlit
pandas
matplotlib

